requirements.txt
numpy
opencv-python
onnxruntime-gpu
pillow
pyyaml
matplotlib